Skip returning *Path for path commands; refactor SFNT
