+++ examples/graph/main.go
@@ -103,11 +103,13 @@ func draw(c *canvas.Context) {
 
 	frame := canvas.Rectangle(xmax-xmin, ymax-ymin).Translate(xmin, ymin)
 	for x := 10.0 * float64(int(xmin/10.0)+1); x < xmax; x += 10.0 {
-		frame.MoveTo(x, ymin).LineTo(x, ymin+2.0/yscale)
+		frame.MoveTo(x, ymin)
+		frame.LineTo(x, ymin+2.0/yscale)
 		c.DrawText(x, ymin-tickFace.Metrics().LineHeight/yscale, canvas.NewTextLine(tickFace, fmt.Sprintf("%g", x), canvas.Center))
 	}
 	for y := 10.0 * float64(int(ymin/10.0)+1); y < ymax; y += 10.0 {
-		frame.MoveTo(xmin, y).LineTo(xmin+2.0/xscale, y)
+		frame.MoveTo(xmin, y)
+		frame.LineTo(xmin+2.0/xscale, y)
 		c.DrawText(xmin, y-(tickFace.Metrics().CapHeight/2.0)/yscale, canvas.NewTextLine(tickFace, fmt.Sprintf("%g ", y), canvas.Right))
 	}
 	c.DrawPath(0.0, 0.0, frame.Transform(viewport))
